+++ R/index-methods.R
@@ -61,13 +61,14 @@ as.double.index <- function(x, ...) {
 }
 
 `[[<-.index` <- function(x, i, j, value) {
+  # make a slice matrix that indexes the same as if 'x' were a matrix
   dim <- c(length(x$levels), length(x$time))
   row_slice <- .row(dim)
   col_slice <- .col(dim)
   dimnames(row_slice) <- dimnames(col_slice) <- list(x$levels, x$time)
-
   i <- row_slice[[i, j]]
   j <- col_slice[[i, j]]
+  
   level <- x$levels[[i]]
   period <- x$time[[j]]
 
@@ -101,17 +102,17 @@ head.index <- function(x, n = 6L, ...) {
   nl <- levels <- length(x$levels)
   np <- periods <- length(x$time)
   if (!is.na(n[1L])) {
-    nl <- if (n[1L] < 0L) {
-      max(levels + n[1L], 0L)
+    if (n[1L] < 0L) {
+      nl <- max(levels + n[1L], 0L)
     } else {
-      min(n[1L], levels)
+      nl <- min(n[1L], levels)
     }
   }
   if (!is.na(n[2L])) {
-    np <- if (n[2L] < 0L) {
-      max(periods + n[2L], 0L)
+    if (n[2L] < 0L) {
+      np <- max(periods + n[2L], 0L)
     } else {
-      min(n[2L], periods)
+      np <- min(n[2L], periods)
     }
   }
   x[seq_len(nl), seq_len(np)]
@@ -121,17 +122,17 @@ tail.index <- function(x, n = 6L, ...) {
   nl <- levels <- length(x$levels)
   np <- periods <- length(x$time)
   if (!is.na(n[1L])) {
-    nl <- if (n[1L] < 0L) {
-      max(levels + n[1L], 0L)
+    if (n[1L] < 0L) {
+      nl <- max(levels + n[1L], 0L)
     } else {
-      min(n[1L], levels)
+      nl <- min(n[1L], levels)
     }
   }
   if (!is.na(n[2L])) {
-    np <- if (n[2L] < 0L) {
-      max(periods + n[2L], 0L)
+    if (n[2L] < 0L) {
+      np <- max(periods + n[2L], 0L)
     } else {
-      min(n[2L], periods)
+      np <- min(n[2L], periods)
     }
   }
   i <- seq.int(to = levels, length.out = nl)